split down templateHelpers
