+++ packages/desktop-client/src/components/modals/BudgetAutomationsModal/EmptyState.tsx
@@ -7,7 +7,7 @@ import { View } from '@actual-app/components/view';
 
 import { getAutomationExamples } from '#components/budget/goals/automationExamples';
 import type { AutomationEntry } from '#components/budget/goals/automationExamples';
-import { getDisplayTemplateMeta } from '#components/budget/goals/templateHelpers';
+import { getDisplayTemplateMeta } from '#components/budget/goals/displayTemplateMeta';
 
 type EmptyStateProps = {
   onAdd: (create: () => AutomationEntry) => void;
